kiosk-agent/requirements.txt
# OpenAI SDK for LLM integration
openai>=1.12.0

# Mem0ai for memory management
mem0ai>=0.0.10

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.9.0

# Optional: Additional utilities
tenacity>=8.2.0